+++ pass-deposit-services/deposit-cri/src/main/java/org/eclipse/pass/deposit/cri/CriticalRepositoryInteraction.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.eclipse.pass.support.messaging.cri;
+package org.eclipse.pass.deposit.cri;
 
 import java.util.Optional;
 import java.util.function.BiPredicate;
