+++ pass-deposit-services/deposit-messaging/src/main/java/org/eclipse/pass/deposit/cri/CriticalPath.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.eclipse.pass.support.messaging.cri;
+package org.eclipse.pass.deposit.cri;
 
 import java.util.function.BiPredicate;
 import java.util.function.Function;
