+++ pass-deposit-services/deposit-messaging/src/main/java/org/eclipse/pass/deposit/messaging/DepositServiceErrorHandler.java
@@ -16,7 +16,7 @@
 package org.eclipse.pass.deposit.messaging;
 
 import org.eclipse.pass.deposit.messaging.service.DepositUtil;
-import org.eclipse.pass.support.messaging.cri.CriticalRepositoryInteraction;
+import org.eclipse.pass.deposit.cri.CriticalRepositoryInteraction;
 import org.eclipse.pass.support.client.model.Deposit;
 import org.eclipse.pass.support.client.model.Submission;
 import org.slf4j.Logger;
