+++ pass-deposit-services/deposit-messaging/src/main/java/org/eclipse/pass/deposit/messaging/runner/SubmittedUpdateRunner.java
@@ -15,7 +15,7 @@
  */
 package org.eclipse.pass.deposit.messaging.runner;
 
-import static org.eclipse.pass.support.messaging.constants.Constants.PassEntity;
+import static org.eclipse.pass.deposit.messaging.Constants.PassEntity;
 
 import java.io.IOException;
 import java.util.Collection;
