+++ pass-deposit-services/deposit-messaging/src/main/java/org/eclipse/pass/deposit/messaging/service/DepositUtil.java
@@ -28,8 +28,8 @@ import org.eclipse.pass.deposit.messaging.policy.TerminalSubmissionStatusPolicy;
 import org.eclipse.pass.deposit.messaging.status.DepositStatusEvaluator;
 import org.eclipse.pass.deposit.messaging.status.SubmissionStatusEvaluator;
 import org.eclipse.pass.deposit.model.DepositSubmission;
-import org.eclipse.pass.support.messaging.cri.CriticalRepositoryInteraction;
-import org.eclipse.pass.support.messaging.cri.CriticalRepositoryInteraction.CriticalResult;
+import org.eclipse.pass.deposit.cri.CriticalRepositoryInteraction;
+import org.eclipse.pass.deposit.cri.CriticalRepositoryInteraction.CriticalResult;
 import org.eclipse.pass.support.client.model.AggregatedDepositStatus;
 import org.eclipse.pass.support.client.model.Deposit;
 import org.eclipse.pass.support.client.model.DepositStatus;
