+++ pass-deposit-services/deposit-messaging/src/main/java/org/eclipse/pass/deposit/messaging/service/SubmissionProcessor.java
@@ -36,8 +36,8 @@ import org.eclipse.pass.deposit.messaging.policy.Policy;
 import org.eclipse.pass.deposit.messaging.policy.SubmissionPolicy;
 import org.eclipse.pass.deposit.model.DepositFile;
 import org.eclipse.pass.deposit.model.DepositSubmission;
-import org.eclipse.pass.support.messaging.cri.CriticalRepositoryInteraction;
-import org.eclipse.pass.support.messaging.cri.CriticalRepositoryInteraction.CriticalResult;
+import org.eclipse.pass.deposit.cri.CriticalRepositoryInteraction;
+import org.eclipse.pass.deposit.cri.CriticalRepositoryInteraction.CriticalResult;
 import org.eclipse.pass.support.client.PassClient;
 import org.eclipse.pass.support.client.model.AggregatedDepositStatus;
 import org.eclipse.pass.support.client.model.Deposit;
