+++ pass-deposit-services/deposit-messaging/src/main/java/org/eclipse/pass/deposit/messaging/status/SwordDspaceDepositStatus.java
@@ -18,7 +18,7 @@ package org.eclipse.pass.deposit.messaging.status;
 import java.net.URI;
 import java.util.stream.Stream;
 
-import org.eclipse.pass.support.messaging.constants.Constants;
+import org.eclipse.pass.deposit.messaging.Constants;
 
 /**
  * Possible states of a deposit to DSpace via SWORD.
